Change the 'resolve_image_location' iteration direction
STRY0010424 - Changed the order in which 'resolve_image_location'
iterates through parent themes so that parent/child theme images
can be overridden consistent with how CSS files are compiled.

This is a port of a customization from 2.6 to 2.8. The original
commit id is: b0a45659a9bde864cda632d8a995f2bb96505bcd