MOOD-1058 umn_clean/javascript: Changed where max content button loads
Max content button was prevented from loading in places where there are no blocks.
We removed the blocks from certain pages in the gradebook, causing the max content
button to not be loaded in those pages. There was a strong desire to keep the
button in these places, so we changed the logic on where the button loads.