MOOD-723 local/course: Course name fix to prevent unnecessary suffix
This corrects a problem where a Twin Cities course would have an
underscore and a digit appended to its short and full names when
a course with the same name had been requested for another campus
for the same semester. This shouldn't have been happening. It should
only have happened when there were multiple Twin Cities courses with
the same name for the same semester.
Also, fixed comment at top of function.