MOOD-1059 atto/plugins: Added border style and BG color to table
This commit adds the option for borders to be added and
styled as well as background colors to be set for tables
in the Atto HTML Editor. For those that desire to restrict
use of borders, the plugin settings allow the admins to restrict
how much freedom their users have with borders.

No build necessary, but to reproduce, navigate to:
lib/editor/atto/plugins/table/yui/src/button/
Run PATH=$PATH:~/local/node-js/node-v0.10.33-linux-x64/bin shifter