MOOD-1094 blocks/lockdownbrowser: Install LockDown Browser 2015051800.
Installed Respondus Lockdown Browser from lockdown_v20150518.zip,
which Alex had attached to the ticket.

In blocks/lockdownbrowser/locklibcfg.php, I changed the
environment-specific settings to dummy values since the BAs manage
those from the UI and we do not want passwords in this git repo.

This version appears to fix the problem that we previously addressed
with MOOD-971 e2b9daddc6604f7e28ca8e6, so we should not need to
reapply that commit.